Start porting code to python3.
In order to do the port, I have to use gi.repository instead of pygtk. So, the
application is using Gtk3. The application runs with python3 with errors:
- AttributeError: 'gtk.gdk.X11Window' object has no attribute 'new_gc'
- Unknown event <enum GDK_2BUTTON_PRESS of type GdkEventType>

Moreover, it doesn't run as intended. The pdf is not displayed until you press a key.
In addition, you must press a key many times in order for a slide to show up or to change,
it seems that the number of the current page is not working properly. I guess it is linked
with the transissions.

The timer looks OK.